# OCR-Spring-Backend-main/swagger/ocr-api.yaml
openapi: 3.0.0
info:
  title: OCR API
  description: API for managing OCR apps
  version: 1.0.0
servers:
  - description: SwaggerHub API Auto Mocking
    url: https://localhost
...
paths:
...
  /api/user/draftentry/upload/images/{id}:
    post:
      tags:
        - Draft Entry Uploads
      summary: Upload images for a draft entry
      description: Allows authenticated users to upload multiple images for a specific draft Entry.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                data:
                  type: string
                  description: JSON string containing image metadata
      responses:
        "200":
          description: Images uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      $ref: "#/components/schemas/DraftImage"
                  message:
                    type: string
        "401":
          description: Unauthorized access
        "404":
          description: Entry not found
        "500":
          description: Internal Server Error
  /api/user/draftentry/upload/reports/{id}:
    post:
      tags:
        - Draft Entry Uploads
      summary: Upload reports for a draft entry
      description: Allows authenticated users to upload multiple reports for a specific draft Entry.
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: array
                  items:
                    type: string
                    format: binary
                data:
                  type: string
                  description: JSON string containing report metadata
      responses:
        "200":
          description: Reports uploaded successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      $ref: "#/components/schemas/DraftReport"
                  message:
                    type: string
        "401":
          description: Unauthorized access
        "404":
          description: Entry not found
        "500":
          description: Internal Server Error
  /api/user/upload/images/{id}:
    post:
      tags:
        - User Upload
      summary: Upload images for teleconsultation entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The object ID of the teleconsultation entry
        - name: _id
          in: header
          required: true
          schema:
            type: string
          description: The header _id after going through authentication process
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: string
                  format: binary
                  description: The image file to upload
                data:
                  type: string
                  description: JSON string containing image metadata
                  $ref: "#/components/schemas/Image"
      responses:
        "200":
          description: "Images Uploaded Successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      type: string
                      description: Image data containing JASON string
                      $ref: "#/components/schemas/imageUploadResponse"
                  message:
                    type: string
                    example: "Images Uploaded Successfully"

        "401":
          description: Unauthorized access
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
        "404":
          description: "Entry Not Found"
  /api/user/upload/reports/{id}:
    post:
      tags:
        - User Upload
      summary: Upload reports for teleconsultation entry
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: The object ID of the teleconsultation entry
        - name: _id
          in: header
          required: true
          schema:
            type: string
          description: The header _id after going through authentication process
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                files:
                  type: string
                  format: binary
                  description: The report file to upload
                data:
                  type: string
                  description: JSON string containing report metadata
                  $ref: "#/components/schemas/Report"
      responses:
        "200":
          description: "Reports Uploaded Successfully"
          content:
            application/json:
              schema:
                type: object
                properties:
                  docs:
                    type: array
                    items:
                      type: string
                      description: Report data containing JASON string
                      $ref: "#/components/schemas/imageUploadResponse"
                  message:
                    type: string
                    example: "reports Uploaded Successfully"

        "401":
          description: Unauthorized access
        "500":
          description: Internal Server Error
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Internal Server Error
                  error:
                    type: string
        "404":
          description: "Entry Not Found"
...
components:
  schemas:
...
    Image:
      type: object
      properties:
        telecon_entry_id:
          type: string
        image_name:
          type: string
        loacation:
          type: string
        lesions_appear:
          type: boolean
          default: false
        annotation:
          type: array
          default: []
        predicted_vat:
          type: string
    Report:
      type: object
      properties:
        telecon_entry_id:
          type: string
        report_name:
          type: string
...
    DraftImage:
      type: object
      properties:
        _id:
          type: string
        filename:
          type: string
        path:
          type: string
        draft_entry:
          type: string
    DraftReport:
      type: object
      properties:
        _id:
          type: string
        filename:
          type: string
        path:
          type: string
        draft_entry:
          type: string
    imageUploadResponse:
      type: object
      properties:
        telecon_entry_id: 
          type: string
          example: "642e422546ac21a667ea0a0d"
        image_name: 
          type: string
          example: "image1.jpeg"
        location: 
          type: string
          example: "Upper labial mucosa"
        clinical_diagnosis: 
          type: string
          example: "Normal"
        lesions_appear: 
          type: boolean
          example: true
        annotation: 
          type: array
          example: []
        predicted_cat: 
          type: string
          example: ""
        _id: 
          type: string
          example: "669095c86782e6762ddd5f35"
        createdAt: 
          type: string
          example: "2024-07-12T02:32:40.650Z"
        updatedAt: 
          type: string
          example: "2024-07-12T02:32:40.650Z"
...